Load ggplot2 to enable independednt knitting
On attempting to knit this file, it failed at the point of calls to
'qplot'. This was probably missed because ggplot2 would have been loaded
into the session at the time it was knitted by other users. It's not hard
to imagine that the package would have been loaded by other .Rmd files in
the book. But if one wants to build this page ALONE, it fails. So I
propose a change wherein we call library on 'ggplot2'.