Update rstudio.Rmd
I don't know how you view this, but I decided t add the code to the demo as the screenshot is a bit blurred to see and reproduce the code. 
Made changes on the key bindings to make them intuitive and consistent with the rest of the other examples you have in the book.